fix(documents-table): No se puede eliminar el documento CERT_RADICA
En la tabla de documentos, ahora el documento de "CERT_RADICA" no se
puede eliminar. El botón de eliminado se deshabilita para este documento
y adicional, si por alguna razon el usuario entra al proceso de
eliminación habrá una notificación que le aparecerá diciendo que no se
puede eliminar este archivo.

+++ src/app/features/bpm-workflows/components/documents-main/documents-table/documents-table.component.ts
@@ -352,8 +352,27 @@ export class DocumentsMainTableComponent implements AfterViewInit, OnInit {
     });
   }
 
+  disableDelete(row: AttachmentCollection): boolean {
+    if (!row || !row.originalFileName) return false;
+    
+    console.log(row.originalFileName);
+
+    return row.originalFileName.startsWith('CERT_RADI');
+  }
+
   onDelete(row: AttachmentCollection): void {
     // Muestra la alerta de confirmación de eliminación
+    if (this.disableDelete(row)) {
+      Swal.fire({
+        text: 'No se puede eliminar este archivo',
+        icon: 'warning',
+        confirmButtonColor: '#3085d6',
+        confirmButtonText: 'Aceptar',
+        timer: 30000
+      });
+      return;
+    }
+
     this.confirmDialog.fire().then((result) => {
       if (result.isConfirmed) {
         this.deleteFile(row);
